// fe/src/app/api/user/model/userResponse.ts
/**
 * Shop App API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponse { 
    id?: number;
    username?: string;
    email?: string;
    phone?: string;
    address?: string;
    provinceCode?: number;
    districtCode?: number;
    wardCode?: number;
    roles?: Array<string>;
    pushToken?: string;
}